ci: cache more Go files (#5140)
Despite the caching that was in place previously, the `go-get-deps`
stage still took quite a while because the caching did not cover
everything that `go install` touches. This change expands the caching to
cover the relevant directories. It also moves the `go install` commands
into separate files so that they can be accounted for in the cache key.

In my testing, this saves 60-90 seconds each time `go-get-deps` is run,
which speeds up several different jobs. Meanwhile, it increases the
cache size from 465 MiB to 676 MiB.