feat: checkpoint download through master (#4989)
Users do not always have access to checkpoint storage, and hence downloading
checkpoints with the CLI does not always work. This feature overcomes the
limitation by supporting checkpoint downloads through the master since t is
often easier to grant access to the master than to individual users.  This
feature only supports AWS S3 for now and relies on implicit IAM permissions on
the master.  In the future, it can be extended to support other checkpoint
storage types and take explicitly provided credentials.

A single endpoint is added:
- `/checkpoints/:checkpoint_uuid/`

The implementation downloads the checkpoint by making API calls to the S3
endpoint, packages it as a tgz or a zip file, and then sends the file to the
client in the HTTP response. The whole process operates as streaming. No
temporary files are stored or cached.