ci: bump cache buster version (#5189)
For some reason, it looks like c82be870, the change introduced by #5140,
hasn't had the time-saving effect it was supposed to. The only
explanation that comes to mind is that I accidentally created an
incomplete cache item while testing that is now being used. Feels
unlikely, but in any case here's a version bump to be sure to get away
from anything like that.