feat: support configuring agent self-shutdown for `det deploy aws` (#5241)
This just threads through a command-line option to `det deploy aws` all
the way down to adding the appropriate config entry to the agent. Since
the hook is configured in a nested object in the config and has a list
value, I didn't want to use another environment variable to configure
it, so the provisioner config now includes a free-form JSON field that
gets inserted into the agent config file.

The new arguments to `docker run` in the agent setup script are
necessary to access and run `shutdown` from within the container.