feat: actually support --device strings
Our docs claim that resources.devices may be the equivalent strings as
one would pass to docker run --device, but that has never been true.

We like what the docs promise, so this PR adds support for exactly that.

We continue to support the map form as well, because devcies with colons
in their name (such as some pci devices) can't be passed in the --device
strings.