fix: exclude allocations without start times from aggregation (#5329)
The allocation aggregation query was designed under the assumption that
everything has a start time, since that was true for trials. At some
point after the query switched to using the allocations table, we added
the PENDING state, within which an allocation has NULL start time.

In PostgreSQL, a NULL start or end for a range indicates that the range
is unbounded in that direction, so an allocation that is in PENDING when
an aggregation runs registers as having been running for the whole day
of the aggregation. We should just ignore those instead.

The case of NULL end time should already be handled, since that has
always been a possibility.