chore: generate PATCH-friendly bindings (#5325)
For classes in bindings.py, avoid defaulting to None via default
parameter values.

Instead, explicitly detect which optional parameters are set, convert
them to None in the `__init__()` body, then keep track of which
parameters were explicitly set.

Then `to_json()`, accepts an `omit_unset` parameter to only emit
parameters which were explicitly set.  This is used for all request
bodies, but the default is False since the most common usage for
`to_json()` is in CLI outputs for --json, where we don't want to change
the existing behavior.

This strategy should play nicely with the FieldMask behavior that
grpc-gateway supports.