refactor: rewrite docker lib without actors [DET-8300] (#4943)
I've taken a lot of care to pull stuff like `aproto.MasterMessage` and
`cproto.ID`'s out of this interface. I figured the better, more
encapsulated our interfaces, the easier stuff would get to test, and to
extend and reuse. So far I feel validated.


I also took a lot of care to design the concurrent parts of the API in
a way that is the most user friendly; APIs mostly take channels instead
of returning them, to let users control back pressure and reuse
channels in their code without much fuss, some APIs return channels,
but only if they are returning a bounded result, and I used directional
channels everywhere, etc. Probably the most important is it is as
un-concurrent as possible so it feels as much like "normal code" as
possible.