refactor: rewrite container lib without actors [DET-8301] (#5384)
The primary goal of this change is refactoring the container actor away
from the actor system. Secondary goals were increased test coverage,
finding good concurrency patterns, and increased locality of code,
meaning the interface for this package should, as much as possible,
not be tightly coupled to those of other packages.