chore: delete try_reauth from authentication logic (#5542)
try_reauth was literally always true, except:
- in a unit test to make sure it worked when it was False
- in logout code, which recently was refactored to not use
  Authentication at all

Furthermore, try_reauth does not make sense to expose to end users.  At
first glance, it sounds like the equivalent of a --noninteractive
commandline flag for batch scripts, but that's only part of what it
does.  try_auth=False has the additional effect that we would not
attempt to log in with the default credentials.  So if we want to give
users a noninteractive setting, that's fine, but it also isn't what
try_reauth does currently.