chore: schemas implement generic-like behavior (#3763)
Remove the type casting that was necessary when the expconf project
landed, but is no longer necessary with go generics.

It turns out that for Copyable, Mergable, and Defaultable, go generics
are not actually helpful.  They are not helpful because the desired
Copyable interface would have to have a constraint of [T Copyable[T]],
which is not a concrete interface, and therefore reflect code can't
check if the type is covertible to such an interface.

Generics could work in another strategy that used no reflect code, so
long as every field in every struct were Copyable/Mergable/Defaultable.
That would involve an annoying amount of type conversions.  Perhaps that
would be acceptable, and the generated getters and setters could do the
annoying conversions for you.  Still it would require most structs to
have Copy methods that simply called .Copy on all of their members.