test: adding tests for Oauth in EE (#582)
Adding e2e_tests for EE in the determined-ee repo. The e2e_tests can't be made generic to for OSS and EE because the master config for EE e2e_tests needs scim which OSS won't recognize. Scim and Oauth is specific to EE. Tests for OSS have also been added as part of the feature PR.

[excluding e2e_tests changes]