fix: Report incorrect configuration message to stderr [FOUNDENG-370] (#5578)
When unable to communicate with the master, trial_prep issues a helpful message, but
it reported to stdout instead of stderr.   Move it to stderr such that it will be highlighted
as an ERROR on HPC clusters where it is particularly a useful message.   It also allows
the Dispatcher RM to drop the display of stdout on job initialization errors.