refactor: rewrite container manager lib without actors (#5386)
The primary goal of this change is refactoring the container
manager away from the actor system. Secondary goals,
repeating again, were clearer interfaces and more testing.
As a tactical maneuver to help some of these goals, I removed
the revalidate flow and plan to always run the reattach flow
instead.