ci: save time by only storing necessary artifacts (#5629)
`master/dist` contains the many individual files from the output of the
docs build, which make the upload take a long time but are unnecessary
to upload at all here (the docs output is uploaded from the docs build
step). We really only want the packages from this step (see #4124).