feat: allow configuring multiple RPs for Kubernetes [DET-8711] (#5698)
The RPs do not correspond to namespaces for the moment; this just sets
up the ability to have multiple logical resource pools at all.

Some of the APIs return slightly incorrect values, since this is still
pushing everything through a single pods actor, which is doing some
summarization without being aware of the multiple RPs. I'm going to call
that okay for this PR, since cluster behavior won't change without
config changes, this is undocumented, and the plan is to do more work
before the next release anyway.

Much of the new code in the RM is copied directly from the agent RM. I
think it may be possible to fully unify them and move all
agent/Kubernetes-specific code into the pools, seeing how both RMs are
pretty much just managing the fleet of RP actors and doing some of
mostly non-cluster-specific bookkeeping, but that's a task for the
future.