chore: Add test-e2e-slurm-podman (#543)
Duplicate the short test-e2e-slurm test to run with
PodMan instead of the default singularity.

Add node cleanup during startup -- to be sure all nodes
are online, and that any bad podman state is cleared.
Define and create XDG_RUNTIME_DIR for use by PodMan.

Drop pytest -s which suppresses capture of output/error messages
in the test results.   This should make triage eaiser.

Change the docker reference of missing.image to be fully-qualified such
that we get a consistent message.  Newer Podman versions will give the error
`Error: short-name resolution enforced but cannot prompt without a TTY` on an
docker reference that is not fully specified.

Enable debugging of pytest results by adding an echo of the pytest output status.
Sometimes we no failure reports when some test cases have failed.
This may help narrow it down when we hit this case.

[e2e_tests changes only]