feat: add det.import_from_path
import_from_path allows users to import arbitrary code from a
checkpoint, even if the modules in the checkpoint have the same name as
modules they already have imported, but contain different code.

This is common when importing, for example, an old model_def.py that has
been updated since the original checkpoint was saved.