remove broken tests
antd v5 has a breaking change which changes how promise rejections are handled
during the action button handler. Instead of swallowing the rejection and
logging it to the console, it now catches and rethrows it to allow the
global unhandled exception handler to fire. However, because the action
button handler's promise isn't returned anywhere, there's no place to
handle it in jest, so tests that relied on the old behavior are broken
and need to be removed.