fix unthemed dialogs
This fixes a bunch of unthemed notifications, modals, and messages that
appeared through the app. This happend due to the upgrade breaking the
behavior of the static methods exported by antdesign. Theme-aware static
methods are provided by the App.useApp hook, so we created a small
module that wraps that while also allowing for static calls. It falls
back to the default static calls for testing purposes