Fix hover color on tables with context menus
This fixes the hover color on tables with context menus. It appears that
antd v5 applies a hover color via pseudoclass, so our code applying it
via manually applying and removing the `ant-table-row-selected` class is
unnecessary. In addition, v5 now highlights selected rows on hover,
causing the hover color issue. There's also a light refactor of some of
the types in the module to get rid of some casts and explicit `any`
assignments.