fix: report configured slot type in Kubernetes
In the recent refactors, the line that reported the slot type from the
config got removed accidentally, causing all slots to reported as
unspecified type.

Ideally, the values should actually be read on a per-RP basis using
quotas or something, but that is still being figured out, so let's at
least get this value in here first.