style: Update antd to v5 (#5788)
* Update antd to v5

This updates antd to v5.

current status:

[x] Remove build dependencies stemming from the previous version's dependency on momentjs
and less.

[x] Reimplement dark/light style switching

[ ] Categorize and possibly fix style breakages stemming from update

* fix some tests

* fix spinner test

* remove broken tests

antd v5 has a breaking change which changes how promise rejections are handled
during the action button handler. Instead of swallowing the rejection and
logging it to the console, it now catches and rethrows it to allow the
global unhandled exception handler to fire. However, because the action
button handler's promise isn't returned anywhere, there's no place to
handle it in jest, so tests that relied on the old behavior are broken
and need to be removed.

* fix create experiment modal

The create experiment modal had a lot of styles that didn't apply to it
even before the v4-v5 upgrade. This removes those while also improving
the width of the modal body for modals that don't use the antd native
icon functionality

* fix hyperparameter search modal

* remove extra padding on modal bodies

* unset default typography font size

* Fix remaining style issues

This fixes input controls being darker than the background, the change
to border radii, and the experiment page overflowing

* Fix input color background

* fix: actionable button effect alignment

* fix unthemed dialogs

This fixes a bunch of unthemed notifications, modals, and messages that
appeared through the app. This happend due to the upgrade breaking the
behavior of the static methods exported by antdesign. Theme-aware static
methods are provided by the App.useApp hook, so we created a small
module that wraps that while also allowing for static calls. It falls
back to the default static calls for testing purposes

* Fix hover color on tables with context menus

This fixes the hover color on tables with context menus. It appears that
antd v5 applies a hover color via pseudoclass, so our code applying it
via manually applying and removing the `ant-table-row-selected` class is
unnecessary. In addition, v5 now highlights selected rows on hover,
causing the hover color issue. There's also a light refactor of some of
the types in the module to get rid of some casts and explicit `any`
assignments.

* force spinner test to work in jsdom

* remove unnecessary jest enviroment declaration

* ensure form control backgrounds are transparent

* fix remaining style issues

* remove extraneous usetheme

* fix fouc issues

The css-in-js solution used by antd has a weird race condition when
extracting css from tokens. This patches over it with a small timeout

---------

Co-authored-by: Keita Nonaka <keita.nonaka@hpe.com>