test: tweak Elasticsearch logging tests
The Elasticsearch agent logging tests have been very sporadically flaking for a while. I tracked down one of the failures (the one in `fluent_api_test.go`) to Fluent Bit not receiving the last log line; I'm not sure why that's happening, but it might possibly be related to the last line being partial. I don't really know what's going on, and I suppose partial lines should work anyway, but it seems like enough of an edge case that I'm just making the last line partial, and also increasing the amount of logging in case this still fails in the future. The test also now gets a bounded number of tries, so that we can see output on failure instead of the test runner just getting killed.

The other test fails with the lines being out of order instead, which is even stranger, but I haven't reproduced it enough to know more; it's getting a similar treatment, and now we'll hope for the best.