refactor: change auth store to use observables
This changes the auth store to use observables. Instead of exporting an
object containing the observables and methods to interact with them, the
read-only observables and modification methods are exported directly
from the module. This means that retrieving the accessors can happen in
the import statement, but renaming variables can be a bit of a pain. In
addition, a common use case for the auth loadable was to get the
`isAuthenticated` property through a match statement. Because these
calls were pretty much the same everywhere, that pattern was lifted to a select.