feat: max-length-based searcher API
Every training loop I know of (except ours) is based on a finite maximum
length, with possible early stopping.

It would be nice if our searchers also behaved the same way.  It turns
out that we can basically shim our existing searchers into this
behavior, and pass the "searcher-decided-stop" signal through the
should_preempt signal, rather than having it be separate.

Long-term, it would be nice if searchers examined validation metrics
rather than requiring explicit report_metrics() calls.  It would also be
nice if progress were not reported through the searcher... although it
might be hard to calculate progress if there's no single "brain" with an
endstate goal for the experiment to know how much total work is
supposed to be done.  Perhaps, experiments should have their own
progress that is separate from a single tasks' progress?

This API might need more thought before we jump ship on our current API.
Not because I like our current API, but because jumping ship is
expensive and we should be sure when we do it.