test: switch test-e2e-cpu-double to use priority scheduler
One of the tests in the changed run flakes occasionally due to a bug in
the fair share scheduler that we're not planning to address
soon (DET-8883), so let's try changing over to priority.

We could also change that test to a managed devcluster test, but that
would be more work, and also this way I think we get a better overall
split of testing between the two schedulers.