fix: Persist theme settings (#5954)
This fixes an issue where the user's theme preference wasn't respected
on refreshing the page. This was because the theme context was moved
above the settings context as part of the antd5 upgrade. That's now
fixed, but there is a minor chance that some dialogues that appear at
the same time as the initial render will not be aware of the theme.