ci: retry Go package download
We had one CI failure [1] caused by the Go package download somehow
fetching some incorrect file. There's no way to know now exactly what it
was, but it seems like it was probably a 5xx status page or something,
and the issue was evidently fairly transient, since it only happened
that once. So now we tell curl to retry that download a bunch to try to
get past that happening again.

Also upgrade from 1.18 to 1.18.10 while we're here.

[1] https://app.circleci.com/pipelines/github/determined-ai/determined/33339/workflows/dd7d6e62-e6c8-4152-bb41-3d313ae9adab/jobs/1170863