build: Solve typescript library checks (#5984)
This makes a few changes to the codebase to ensure that our typescript
checks use the library definitions when checking. We update a few
DefinitelyTyped types and remove a few types that weren't in use. The
InteractiveTable component is modified slightly because the column
resize props overlapped with the resize props the antd table type gets
from the dom types. We removed the dayjs declaration because it's no
longer necessary, and fixed the notebookjs type declaration.

the largest part of this pr is lightly refactoring the typescript code
generated from swagger-codegen to not use portable-fetch and also not
directly delete the querystring when generating urls in order to fix
some oddness stemming from url object behavior. In order to do so, we
compile a class that extends the preexisting typescript codegen and
ensure it's in the classpath when we generate the types.