fix: handle nans in GET /api/v1/experiments/:id?showTrialData=true (#5998)
In the `show_trial_data=true` feature, we present the `best_trial` as a sub-struct
of an experiment using `json_build_object` in Postgres. This doesn't work out, because
when Postgres jsonifies floats, it turns `inf` and `nan` into true strings and golang ends
up trying to scan the string "Infinity" into a `float64`. We fix this by avoiding going
through `json` in Postgres.