fix: guard against store polling when not logged in (#6018)
* fix: guard against store polling when not logged in

This fixes an issue where a login loop could occur when using Okta. Okta
login requires auth to read the token after mounting the app. The
initial call to check the auth happens in the App component. When we
moved the agent and resource pool calls to an observable store, we moved
the calls to init the information into a polling funciton called above
the App component, causing a race condition where the auth would
complete but the user would be kicked back to the login screen
because the initial polling calls made before the user could be
authenticated failed. To be safe, we fix a similar pattern happening in
the navigation component which causes a failed request on the login page.

* wrap polling function in useCallback

(cherry picked from commit 67d1a3ccc847d7a4566114b8fe2671101f0c38d5)