build: [MLG-305] Check for supported go version at make. (#5975)
Determined cannot be currently built with go1.20. This adds a target in
the Makefile that checks the system go version against the supported go
version and errors if the system go is not supported.

This is just a simple fix for an existing problem. It temporizes any
decisions about more generally checking requirements, and also leaves
for a future commit updates to the library to make determined buildable
with more recent go.

This was tested by hand against go1.18.10, go1.19.5, and go1.20.