fix: Drop incorrect max_slots support code [FOUNDENG-454] (#691)
The expconf resources.max_slots attribute is documented as managed by
Slurm, but there was a partial (non-functional) implementation in
DispatchRM as well.  If we ever submitted max_slots worth of work
the remainder would remain QUEUED forever.   I did attempt to fix
the existing support, but the obvious fixes were not sufficient.  Since
this is already documented as managed by Slurm, just remove the support
and let the workload managers manage it.