fix: Avoid startup lag due to debug mode cleanup [FOUNDENG-477] (#707)
DispatcherRM debug mode defers cleanup of dispatches to enable access to the detailed
job logs to enable triage when things are not working.   Because it was intended for internal debugging
only, it had not been optimized and did the cleanup of the dispatches synchronously.  This blocks
startup of the Determined master.   Enhance the debug mode cleanup to be done in a go routine, and
only cleanup dispatches associated with completed allocation (previously it cleaned up
everything and forced use of a restart).   This change should make it reasonable for
use on a customer site when necessary.