fix: Handle subdirs in expconf model_def with Launcher [FOUNDENG-479] (#709)
If a model_def directory contained a subdir, it was incorrectly treated
as needing a softlink (but that is needed only for the top-level
child dirs of /run/determined) which causes a failure in the launcher
when trying to expand the archive.  Only convert top-level paths
under /run/determined to softlinks.