ci: send Slack notification on GKE node pool creation failure (#6152)
In order to prevent quota failures from showing up as CI failures, this
makes node pool creation failure send a Slack notification and mark the
job as successful.

I couldn't figure out how to use the Slack orb while distinguishing this
particular situation from the general failure case, so I just slapped in
a direct request to the already-configured Slack webhook for sending
messages to #ci-bots.

`circleci-agent step halt` marks the job as successful, which is why we
want a notification at all. For some reason, CircleCI fails to provide
an equivalent for marking the job as canceled or some other state
besides success/failure; we could make a call to the CircleCI API to
cancel the current job, but that would rely on having a CircleCI token
available, which we're trying to get away from.