fix: Close expiriment fork/ continue trial modal properly
This fixes an issue where the experiment fork/continue trial modals
would not open again after the user closed the modal via the close
button. This is because we overloaded the modal's close handler to
toggle switching to the config view. This functionality is replaced by
simply overriding the onClick prop on the cancel button.