chore: move assigned event to correct location (#6197) [DET-9012]
The assigned event is sent correctly for the agent rm, but (confusingly)
when the 'resources' are allocated in slurm/k8s, it is just a handle to submit
a slurm job/k8s pod, so the CLI can print "something something assigned
to an agent" when the task has only been dropped onto the underlying
RM's queue. This moves the message (to where allocation start time is set,
for the same reason) so it is less confusing.