apply fixes for vitest and jest-dom matchers
This fixes an issue where the globals instated by vitest clashed with
the jest globals that were required by @testing-library/jest-dom. We
had to exclude the jest types from being imported by overriding the
resolved version of the jest types to a nonexistent folder and also
patching the typings of the jest-dom package such that it no longer
tried to import the jest types.