ci: increase timeout and retries for `pip install`
We keep seeing timeouts on these installs, so increase to a timeout of
30 seconds with 10 retries, up from the default of 15 seconds and 5
retries.

There are other places that run `pip install` in CI, but I think these
are the only ones I've seen fail (presumably because all the really big
installs go through them), so let's stick with them for now.