fix: Restore proper SIGTERM handling for Slurm/PBS [DET-9124]
In Slurm/PBS SIGTERM is reserved for notification of a pending preemption, so handle it in launch.py instead of
forwarding to the children processes. This change exempts det.util.forward_signals from forwarding SIGTERM when
running on Slurm/PBS such that preemption can be triggered by the chief rank.