chore: add e2e test for preemption on hpc cluster (FOUNDENG-497) (#726)
Add an e2e test for preemption on an HPC cluster.
Preemption:
When users launch a set of experiments requesting different levels of priorities and resources, and when there are a limited set of resources, high-priority experiments can cancel or re-queue low-priority experiments.
Preemption is dependent upon the underlying HPC system and the WLM (SLURM/PBS) setup. Nodes in an HPC system are typically divided into multiple partitions (a logical grouping of nodes into possibly overlapping sets) used for different purposes. Using WLMs sysadmins typically assign varying levels of priority for each partition. Also, users can request the WLM to provide a specific partition and priority level for their job.
In this PR I added an e2e test for the preemption scenario.
In the preemption test case, we launch the two experiments cifar10_pytorch_cancellable and cifar10_pytorch_high_priority in order. Ensure that the cifar10_pytorch_cancellable experiment is re-queued, and the cifar10_pytorch_high_priority experiment runs to completion. After that, the cifar10_pytorch_cancellable experiment is resumed and it runs to completion.
NB: The clusters casablanca-login and znode have one node (8-GPUs) being used as two partitions:
1. defq_GPU_cancellable - partition for low priority and jobs are re-queued if necessary
2. defq_GPU_hipri - partition for high-priority non-cancellable jobs

[e2e_tests changes only]