fix: avoid explosion of prom series per actor with better labelling (#6280)
Didn't quite hit the mark with how I chose the to/from labels for actors - it
was too easy to miss things. This isn't as quite as pretty as actors addrs
but it makes sure we don't oom prom. We make a lot of actors and at least
the types in our system always have a static upper bound.