feat: load PyTorchTrials with kwargs
There wasn't a good option for supporting easy loading of checkpoints
with Trainer API down the "instantiate your own Trial" road. Also, the
whole concept of "instantiate your own Trial" during checkpoint export
was dubious... it is likely nobody actually wanted their Trial instance,
just their model.

There were some promising options with the potential adding of methods
save_model_weights() and load_model_weights() to the PyTorchTrial,
but it was unfortunate that those would have been effectively mandatory
for all Trainer users.

Expanding the existing load_trial_from_checkpoint_path is an option
that is easy for current and most future users, while not precluding us
from offering save_model_weights/load_model_weights in the future.

However, one challenge presented by this interface was that if the user
defines their Trial class directly in their entrypoint script, we aren't
guaranteed to be able to import it in an automatic way.  To work around
this, we rely primarily on some huristcis to import it in most cases,
and also we left the CheckpointLoadContext in the library but
undocumented as an emergency escape hatch if this comes up in
production.