fix: pytorch loading accepts trial_class kwarg
Previously, load_trial_from_checkpoint_path expected to be able to
automatically import the user's Trial class.

Before Trainer, this was always true.  Since Trainer, it is possible to
define a Trial inside the __main__ module, which is not reliably
importable.

In fec6cf23 I attempted to fix this by guessing the path at which
__main__ would be importable.  That strategy fails miserably when inside
a notebook, where __main__.__file__ is not even defined.

Now, the new approach is to simply have the user pass us the Trial class
in cases where we can't import it automatically.

Alternate approaches which were considered:

- Pass the CheckpointLoadContext the path to the checkpoint, and load
  state dicts as context.wrap_*() was called.  This was rejected because
  it changes the timing of when state is loaded between training (after
  __init__) and checkpoint export (during __init__).

- Introduce CheckpointLoadContext.load(path), or some other loading
  function.  This was rejected because it was more verbose and more of a
  departure from th existing load_trial_from_checkpoint_path without
  offering much apparent benefit.