chore: Fix router/app circular dependency (#6370)
Some cleanup ahead of WEB-1043 where we need the ability to display the
designkit without the server running -- this removes the circular
dependency where the router is dependent on itself. Instead, we wrap the
router in an object that's instantiated in the index file. The mock is
also updated for testing purposes.