build: fix examples makefile
There was a bug causing examples to not rebuild when they should have,
which was causing issues for engineers unaware of the bug.
Specifically, when new files appeared in glob patterns, but which were
created before the build.stamp (which can happen if you
build-then-rebase), then you could run `make -C examples build` and it
would insist there was no work to do.  The same issue would happen if a
file was deleted, no rebasing required.

Build bugs are always much worse for anyone not intimately familiar with
the build system, and as we grow this is getting more obvious.

The new logic is a bit faster for clean builds (1.5s to 1.1s for me) and
a bit slower on noop builds (0.05s to 0.11s).

Additionally, the new logic makes a couple of strategic sacrifices:

- The per-example incremental rebuild is gone.  The pattern is now
  "rebuild all examples if any change is detected".  This strategy is
  more performant and results in a simpler makefile.  In reality, I
  don't think anybody needs incremental builds of experiment tarballs;
  either somebody is writing experiments or they are building docs, but
  rarely are both happening in one workflow.

- The examples/Makefile is more complicated than if we just did a
  rebuild-from-scratch strategy every time.  However, for non-parallel
  builds, this strategy is nearly 10x faster than rebuilding each time.
  And there's no real reason to rebuild every example every time before
  building the docs (which would effectively result).