test: cleanup should kill hung experiments, not cancel (#6484) [DET-9284]
In slurm-world, there was a bug where we never thought trials were ready so
we skipped graceful cancellation and killed them, always. This was recently
fixed but since there is also an issue involving a hanging experiment in tests
and test code cancels, never kills, experiments that won't complete hang
forever in tests. This change just updates the testing code to kill not cancel
experiments that are stuck.