fix: Show container started messages in exp log [DET-9285] (#796)
Recursion is seeing occasional delayed start of containers on specific nodes.
We have messages about container startup, but they are in the master log.
This moves them into the exp log where they are visible and easily found
to assist in identifying problematic container startup.  They are shown
only when there are mulitple containers involved.