fix: Avoid excessive dispatches in debugging mode [FOUNDENG-567] (#817)
When the master is in debug mode we leave dispatches on the disk for triage purposes.
This change adds a periodic cleanup every 18hr in debug mode using the existing
inactive dispatch cleanup method.   We do one immediate cleanup on startup then
if in debug mode, we continue to periodically repeat the cleanup.