docs: stop relying on algolia crawler (#6584)
The Algolia crawler is complex and poorly documented and runs on
infrastructure we don't want in the critical path for a release.

Instead, build our docs to XML, write a python script that scrapes the
XML and pushes records directly into the Algolia index.

This helps us solve need to "stealth-publish" the docs in order to have
a crawler-generated index ready as soon as the release is cut, because
now we can scrape/upload literally any time we want.  Also it solves the
O(N) cralwer runtime problem because scrape/upload takes only a few
seconds and doesn't affect any old outputs.

Lastly, this lets us simplify our docs javascript code a bunch, because
we can upload partial urls instead of complete ones, making our
transformItems function trivial.