ci: fully quarantine `test-det-deploy-local` and `test-stress`
Bit of a heavy hammer, but the issue with these tests appears to not be
specific to any particular test: after some specific tests in
`test-det-deploy-local` were quarantined, other tests that were left
behind started failing in the same way. And now `test-stress` seems to
be showing the same thing as well.