ci: auto-scrape algolia records (#6620)
Scrape + upload Aloglia records of the "dev" version on every commit
that lands on main.  The dev index is used by source builds of the docs.
This implies that old source build won't have a functional search bar,
but that seems fine.

Scrape + upload Alogila records for the official version on every commit
that is either tagged as a release or a release candidate.  There's no
point in keeping extra release-candidate search indices.