ci: reduce overly long CloudFormation stack name
With a prefix of `mmdetection-quarantine`, an S3 bucket in the stack
would try to have a name longer than the maximum of 63 characters, so
stack creation would always fail.

The original `mmdetection` doesn't need to be shortened to work, of
course, but I figure both of them ought to be the same.