chore: refactor columns api (#6563)
This refactors the columns api. The generalcolumn enum values were not
in use for the upcoming work on sorting and filtering because those use
a dsl, so those are replaced by plain text that matches columns in the
experiment. We also include human readable names for the general columns
to use in menus and tables as well as the type of data the column
represents. Because most of the use cases of the columns api are to show
an ungrouped list of options to a user, the output was simplified to an
array, while also keeping the information as to which object the column
relatest to in the `location` field.