test: Add visual diff test to designkit
This adds a visual diff test to designkit. We generate screenshots of
each section of the designkit on the branch in question using playwright
and then again on the main branch. We then use playwright's built in
image comparison facilities to generate a diff between the screenshots.
If differences are found, both screenshots and the diff are output to a
flat html file that displays the results and CI will report in github.