chore: print messages while `det deploy aws (up|down)` waits (#6838)
Sometimes, CloudFormation is slow and a job gets killed by CircleCI
while waiting for the cluster to come up because there's no output for a
long time.

There was evidence that the slowness is actually due to an AWS bug and
that the stack might _never_ come up in those cases, but at least this
way we can find out for sure. And I suppose it's a little bit nicer for
actual users as well.

This adds the output to stack deletion too, for symmetry and the user
experience reason, though I don't remember having any hangs there.

It would not be too hard technically to hack the Boto3 waiter to print
periodic messages without the need for a separate thread, but it didn't
look possible with only the _documented_ interfaces, so threading it is.