fix: trials created by paused experiments should be paused [DET-9493]
NewTrial was hardcoded to state:ActiveState even if the trial.state
was PAUSED, so depending upon timing new trials may show as Active/Queued
when the experiment is actually paused.  Pass the initial state to
NewTrial such that it can be persisted in the DB matching the actual
state in memory.