chore: enable exhaustiveness checks in switch statements
This turns on the typescript-eslint rule that ensures the type of the
expression in a switch statement is `never` after all case statements.
There were three places where we didn't fully cover all cases, which are
fixed here. In addition, we also disable the `no-fallthrough` rule,
because it can sometimes cause a false positive where the fallthrough is
unreachable -- it also overlaps in purpose with the
`noFallthroughCasesInSwitch` typescript config option.