ci: handle cherry-pick marker label in release automation
Following on from #6867, this adds some more steps to the release
automation. The workflows now handle a special label that marks a PR as
a fix to be cherry-picked for the current release. There's also
scaffolding for some future steps for actually handling the
cherry-picks; some of that implementation is left stubbed out for now,
but I believe this should cover the full shape of the Actions lifecycle
we have in mind. Remaining work will lie in filling in that
implementation and updating rph to handle the human side of things.