fix: Improve error message formatting with synchronous launch [FE-5] (#862)
After converting from LaunchAsync to synchronous Launch, we now have
a different path for processing failed launches.   Previously this was
handled by the dispatch_monitor, but now we get back the error as a direct
REST call response.  So handle BadRequest and InternalServerError directly
from the detail message from the response body (if available), instead of
showing the whole error response in JSON.

All the Slurm failure tests still succeeded, but the messages were surrounded
by the additional JSON fields which made it much less appealing.