chore: consolidate and simplify isort config files
The exact duplicates of the config file seem unnecessary, and we can
make use of isort's profile feature to avoid specifying several of the
other options.

The imports of DeepSpeed move because isort now (correctly) sees it as a
third-party package (the determination appears to be based on the
directory containing the config file). Similarly, some other packages no
longer need to be specified.