test: change filter for rbac tests.
Previously, rbac tests were prevented from running if the system found
the string "Unimplemented" when an rbac command was run from the CLI.
Generally, this happens only on the OSS version. But because rbac can be
implemented in EE but not enabled, these tests are now additionally
gated on whether rbac is enabled.