ci: handle cherry-picking PRs for release (#7035)
When a PR with the appropriate label is merged during a release (or a
previously merged PR has the label added), the release branch should now
automatically be updated with the cherry-picked version if possible. If
there is a conflict, the PR will be added to the conflict column in the
current release project for manual handling.