fix: master returns rbacEnabled = false; change filter for rbac tests (#7100)
* Change filter for rbac tests.

Previously, rbac tests were prevented from running if the system found
the string "Unimplemented" when an rbac command was run from the CLI.
Generally, this happens only on the OSS version. But because rbac can be
implemented in EE but not enabled, these tests are now additionally
gated on whether rbac is enabled.

* Add e2e_cpu_rbac mark to tests in test_groups.py

* Add e2e_cpu_rbac as marks in both the pytest.ini and conftest.py

* OSS is never rbacEnabled.

master now says that rbacEnabled is false in OSS no matter its configuration.

* move some utility functions into utils modules

Test files in the tests directory import utility functions from each
other all over the place. This represents a minimal change to make it
possible to remove test_groups.py and test_rbac.py, in case we want to
move them out of OSS into EE.