chore(go): add a generic queue as a drop-in actor inbox replacement (#6962)
Add a generic queue as a drop-in replacement for an actor inboxes when
removing the actors from the system. This is useful to avoid side affects where
switching to a channel would introduce back-pressure. In some cases, it is
difficult to vet such back-pressure for unintended side effects. In others,
it is impossible to add without a much larger change.