test: Add a unit test for direct downloading from S3. (#7174)
I'm not super happy with the state of unit testing for downloading. This
PR takes a random stab at a single code path in a test that lies
somewhere between integration and unit. But downloading has a very broad
interface, and I don't see any unit tests for the storage managers, so I
think this PR represents at least a small improvement.

* New functional test of s3 downloading
-    adds moto to testing requirements to achieve this

* New tests with Mock storage managers
-   Ensures the storage managers are called in test_checkpoint