Fix: use user id for new explist user filter
This fixes an issue where the user filter in the new experiment list
would sometimes fail to find matching experiments for users with display
names set. This is because the webui would send the username, but the
api was expecting the display name or username if set. To fix this, we
update the frontend and backend to use the user id instead.