fix: SDK methods don't update local on failure
Local objects contain copies of data that exists also in the remote
master. These copies can get out of sync.

This change addresses methods that change local objects' data and post
(or patch) a request to master to update the remote copies, too.

Before this change errors over the API would still change local
versions. With this change, we now treat APIErrors as total failures,
and no longer update local copies.

This PR also adds tests for exactly this behavior in all the "update"
methods in the SDK.